requirements.txt
Flask~=2.1.1
hazelcast-python-client
python-consul